+++ AndroidLib/src/test/kotlin/com/bloomberg/selekt/android/support/SupportSQLiteOpenHelperTest.kt
@@ -18,12 +18,12 @@ package com.bloomberg.selekt.android.support
 
 import com.bloomberg.selekt.android.ISQLiteOpenHelper
 import com.bloomberg.selekt.android.SQLiteDatabase
-import com.nhaarman.mockitokotlin2.doReturn
-import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.times
-import com.nhaarman.mockitokotlin2.verify
-import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
-import com.nhaarman.mockitokotlin2.whenever
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.times
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.verifyNoMoreInteractions
+import org.mockito.kotlin.whenever
 import org.junit.Test
 import kotlin.test.assertSame