+++ AndroidLib/src/test/kotlin/com/bloomberg/selekt/android/support/SupportSQLiteProgramKtTest.kt
@@ -17,10 +17,10 @@
 package com.bloomberg.selekt.android.support
 
 import com.bloomberg.selekt.ISQLProgram
-import com.nhaarman.mockitokotlin2.eq
-import com.nhaarman.mockitokotlin2.same
-import com.nhaarman.mockitokotlin2.times
-import com.nhaarman.mockitokotlin2.verify
+import org.mockito.kotlin.eq
+import org.mockito.kotlin.same
+import org.mockito.kotlin.times
+import org.mockito.kotlin.verify
 import org.junit.Before
 import org.junit.Test
 import org.mockito.Mock
@@ -32,7 +32,7 @@ internal class SupportSQLiteProgramKtTest {
 
     @Before
     fun setUp() {
-        MockitoAnnotations.initMocks(this)
+        MockitoAnnotations.openMocks(this)
     }
 
     @Test